chore: upgrade T-Digest library to caio/go-tdigest v5.0.0
Switch from influxdata/tdigest v0.0.1 (2019) to caio/go-tdigest v5.0.0 (2025):

**Why upgrade:**
- influxdata/tdigest: Last release Nov 2019 (v0.0.1)
- caio/go-tdigest: Latest release Nov 2025 (v5.0.0)
- 6 years of improvements, optimizations, and bug fixes
- Actively maintained (updated Jan 2026)
- Better performance and reliability

**API changes:**
- NewWithCompression() → New(tdigest.Compression())
- Add(value, weight) → AddWeighted(value, weight uint64)
- Centroids() → ForEachCentroid(callback)
- Compression field → Compression() method
- Count() returns uint64 instead of float64

**Implementation:**
- Updated import to github.com/caio/go-tdigest/v5
- Adapted decay logic to use ForEachCentroid iterator
- Fixed weight handling (uint64 with rounding)
- Updated all tests to match new API

**Testing:**
- ✅ All AdaptiveMaxCalculator tests pass
- ✅ Build successful
- ✅ Backward compatible behavior (same percentile calculations)

+++ utils/score/adaptive_max_calculator_test.go
@@ -180,7 +180,7 @@ func TestAdaptiveMaxCalculator_Stats(t *testing.T) {
 	require.Greater(t, stats["adaptive_p90"].(float64), 0.0)
 	require.Greater(t, stats["median"].(float64), 0.0)
 	require.Greater(t, stats["p99"].(float64), 0.0)
-	require.Greater(t, stats["total_weight"].(float64), 0.0)
+	require.Greater(t, stats["total_weight"].(uint64), uint64(0)) // Changed from float64 to uint64
 	require.Greater(t, stats["centroid_count"].(int), 0)
 	require.Equal(t, 100.0, stats["compression"].(float64))
 	require.Equal(t, 3600.0, stats["half_life_seconds"].(float64))
@@ -226,7 +226,7 @@ func TestAdaptiveMaxCalculator_LargeDataset(t *testing.T) {
 
 	// Verify stats
 	stats := calc.GetStats()
-	require.Greater(t, stats["total_weight"].(float64), 800.0) // Should have ~900 total weight
+	require.Greater(t, stats["total_weight"].(uint64), uint64(800)) // Should have ~900 total weight
 }
 
 func TestAdaptiveMaxCalculator_Reset(t *testing.T) {
@@ -242,14 +242,14 @@ func TestAdaptiveMaxCalculator_Reset(t *testing.T) {
 
 	// Verify samples were added
 	stats := calc.GetStats()
-	require.Greater(t, stats["total_weight"].(float64), 0.0)
+	require.Greater(t, stats["total_weight"].(uint64), uint64(0))
 
 	// Reset
 	calc.Reset()
 
 	// Verify reset worked
 	stats = calc.GetStats()
-	require.Equal(t, 0.0, stats["total_weight"].(float64))
+	require.Equal(t, uint64(0), stats["total_weight"].(uint64))
 
 	// Can add samples after reset
 	err := calc.AddSample(1.0, time.Now())
